+++ lark/CHANGELOG.md
@@ -4,3 +4,6 @@ v1.0
 
 - `use_accepts` in `UnexpectedInput.match_examples()` is now True by default
 
+- Token priority is now 0 by default
+
+- `v_args(meta=True)` now gives meta as the first argument. i.e. `(meta, children)`